// lib/interfaces/tranzilla.ts
export interface ITranzilla {
    uid: string;
    oper: string;
    Response: string;
    o_cred_type?: string;
    o_tranmode?: string;
    lang?: string;
    expmonth?: string;
    currency?: string;
    o_npay?: string;
    expyear?: string;
    supplier?: string;
    tranmode?: string;
    sum?: string;
    benid?: string;
    ConfirmationCode?: string;
    cardtype?: string;
    cardissuer?: string;
    cardaquirer?: string;
    index?: string;
    Tempref?: string;
    TranzilaTK?: string;
    maxpay?: string;
    createdAt?: Date;
}